+++ src/server.rs
@@ -10,21 +10,20 @@ use std::{
 
 use open62541_sys::{
     UA_NodeId, UA_Server, UA_ServerConfig, UA_Server_addDataSourceVariableNode,
-    UA_Server_addNamespace, UA_Server_deleteNode, UA_Server_getNamespaceByIndex,
-    UA_Server_getNamespaceByName, UA_Server_runUntilInterrupt, __UA_Server_addNode,
-    __UA_Server_write, UA_STATUSCODE_BADNOTFOUND,
+    UA_Server_addNamespace, UA_Server_addReference, UA_Server_deleteNode,
+    UA_Server_deleteReference, UA_Server_getNamespaceByIndex, UA_Server_getNamespaceByName,
+    UA_Server_runUntilInterrupt, __UA_Server_addNode, __UA_Server_write, UA_STATUSCODE_BADNOTFOUND,
 };
 
 use crate::{ua, Attributes, DataType, Error, Result};
 
 pub(crate) use self::node_context::NodeContext;
-use self::node_types::Node;
 pub use self::{
     data_source::{
         DataSource, DataSourceError, DataSourceReadContext, DataSourceResult,
         DataSourceWriteContext,
     },
-    node_types::{ObjectNode, VariableNode},
+    node_types::{Node, ObjectNode, VariableNode},
 };
 
 /// Builder for [`Server`].
@@ -453,6 +452,120 @@ impl Server {
                 ua::NodeId::to_raw_copy(node_id),
                 // Delete all references to this node.
                 true,
+            )
+        });
+        Error::verify_good(&status_code)
+    }
+
+    /// Adds a reference from one node to another.
+    ///
+    /// # Errors
+    ///
+    /// This fails when adding the reference fails.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use open62541::{DataType as _, Node, ServerBuilder, ua};
+    /// # use open62541_sys::{
+    /// #     UA_NS0ID_HASCOMPONENT, UA_NS0ID_OBJECTSFOLDER, UA_NS0ID_ROOTFOLDER, UA_NS0ID_STRING,
+    /// # };
+    /// use open62541_sys::{UA_NS0ID_ORGANIZES};
+    ///
+    /// # #[tokio::main]
+    /// # async fn main() -> anyhow::Result<()> {
+    /// # let (server, _) = ServerBuilder::default().build();
+    /// #
+    /// // let parent_one_node_id = server.add_node(/* snip */)?;
+    /// # let parent_one_node_id = server.add_node(Node::new(
+    /// #     ua::NodeId::ns0(UA_NS0ID_OBJECTSFOLDER),
+    /// #     ua::NodeId::ns0(UA_NS0ID_ORGANIZES),
+    /// #     ua::QualifiedName::new(1, "ParentOne"),
+    /// #     ua::ObjectAttributes::init(),
+    /// # ))?;
+    /// // let parent_two_node_id = server.add_node(/* snip */)?;
+    /// # let parent_two_node_id = server.add_node(Node::new(
+    /// #     ua::NodeId::ns0(UA_NS0ID_OBJECTSFOLDER),
+    /// #     ua::NodeId::ns0(UA_NS0ID_ORGANIZES),
+    /// #     ua::QualifiedName::new(1, "ParentTwo"),
+    /// #     ua::ObjectAttributes::init(),
+    /// # ))?;
+    ///
+    /// let variable_node_id = server.add_node(Node::new(
+    ///     parent_one_node_id.clone(),
+    ///     ua::NodeId::ns0(UA_NS0ID_ORGANIZES),
+    ///     ua::QualifiedName::new(1, "Variable"),
+    ///     ua::VariableAttributes::init(),
+    /// ))?;
+    ///
+    /// // This makes the variable available in two parents.
+    /// server.add_reference(
+    ///     &parent_two_node_id,
+    ///     &ua::NodeId::ns0(UA_NS0ID_ORGANIZES),
+    ///     &variable_node_id.clone().into_expanded_node_id(),
+    ///     true,
+    /// )?;
+    ///
+    /// // Duplicating an existing reference is not allowed.
+    /// let error = server.add_reference(
+    ///     &parent_one_node_id,
+    ///     &ua::NodeId::ns0(UA_NS0ID_ORGANIZES),
+    ///     &variable_node_id.clone().into_expanded_node_id(),
+    ///     true,
+    /// ).unwrap_err();
+    /// assert_eq!(error.status_code(), ua::StatusCode::BADDUPLICATEREFERENCENOTALLOWED);
+    /// #
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn add_reference(
+        &self,
+        source_id: &ua::NodeId,
+        reference_type_id: &ua::NodeId,
+        target_id: &ua::ExpandedNodeId,
+        is_forward: bool,
+    ) -> Result<()> {
+        let status_code = ua::StatusCode::new(unsafe {
+            UA_Server_addReference(
+                // SAFETY: Cast to `mut` pointer, function is marked `UA_THREADSAFE`.
+                self.0.as_ptr().cast_mut(),
+                // SAFETY: The `NodeId` values are used to find internal pointers, are not modified
+                // and no references to these variables exist beyond this function call. Passing by
+                // value is safe here.
+                DataType::to_raw_copy(source_id),
+                DataType::to_raw_copy(reference_type_id),
+                DataType::to_raw_copy(target_id),
+                is_forward,
+            )
+        });
+        Error::verify_good(&status_code)
+    }
+
+    /// Deletes a reference between two nodes.
+    ///
+    /// # Errors
+    ///
+    /// This fails when deleting the reference fails.
+    pub fn delete_reference(
+        &self,
+        source_node_id: &ua::NodeId,
+        reference_type_id: &ua::NodeId,
+        target_node_id: &ua::ExpandedNodeId,
+        is_forward: bool,
+        delete_bidirectional: bool,
+    ) -> Result<()> {
+        let status_code = ua::StatusCode::new(unsafe {
+            UA_Server_deleteReference(
+                // SAFETY: Cast to `mut` pointer, function is marked `UA_THREADSAFE`.
+                self.0.as_ptr().cast_mut(),
+                // SAFETY: The `NodeId` values are used to find internal pointers, are not modified
+                // and no references to these variables exist beyond this function call. Passing by
+                // value is safe here.
+                DataType::to_raw_copy(source_node_id),
+                DataType::to_raw_copy(reference_type_id),
+                is_forward,
+                DataType::to_raw_copy(target_node_id),
+                delete_bidirectional,
             )
         });
         Error::verify_good(&status_code)